requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.1.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
idna==2.8
pip-tools==4.1.0
psycopg2==2.8.3
Pygments==2.4.2
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.3
